refactor: extract trait impls to their respective files
