+++ crates/bevy_auto_plugin_shared/src/__private/type_list.rs
@@ -1,5 +1,6 @@
+use crate::__private::util::generics::{CountGenerics, HasGenericsCollection};
 use darling::{Error, FromMeta};
-use proc_macro2::TokenStream;
+use proc_macro2::{Span, TokenStream};
 use quote::{ToTokens, quote};
 use syn::parse::Parser;
 use syn::{Meta, Token, Type, punctuated::Punctuated};
@@ -97,5 +98,24 @@ mod tests {
         let type_list = &attr.types;
         let tokens = quote! { #type_list };
         assert_eq!(tokens.to_string(), types.to_string());
+    }
+}
+
+impl CountGenerics for TypeList {
+    fn get_span(&self) -> Span {
+        use syn::spanned::Spanned;
+        self.span()
+    }
+
+    fn count_generics(&self) -> syn::Result<usize> {
+        Ok(self.len())
+    }
+}
+
+impl HasGenericsCollection for TypeList {
+    type CollectionItem = Self;
+    type Collection = Vec<Self>;
+    fn generics(&self) -> syn::Result<Self::Collection> {
+        Ok(vec![self.clone()])
     }
 }
